+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## Unreleased
+## 1.0.0
 
 ### Breaking changes
 
 * Remove `DragTargetEvent.x`, `DragTargetEvent.y`, and `DragTargetEvent.offset` (deprecated in `0.85.0`). Use `DragTargetEvent.local_position` for target-relative coordinates or `DragTargetEvent.global_position` for global coordinates ([#6401](https://github.com/flet-dev/flet/pull/6401)) by @ndonkoHenri.
 * Remove `Video.show_controls` (deprecated in `0.85.0`). Set `Video.controls` to `None` to hide controls ([#6463](https://github.com/flet-dev/flet/pull/6463)) by @ndonkoHenri.
 * Remove `Video.playlist_add()` and `Video.playlist_remove()` (deprecated in `0.85.0`). Mutate `Video.playlist` directly with list methods such as `append()` and `pop()` ([#6463](https://github.com/flet-dev/flet/pull/6463)) by @ndonkoHenri.
+* Remove `FletApp.show_app_startup_screen` and `FletApp.app_startup_screen_message` (deprecated in `0.86.0`). Use `FletApp.boot_screen_options` instead, e.g. `boot_screen_options={'spinner_size': 30}` or `boot_screen_options={'startup_message': '...'}` by @ndonkoHenri.
+* Remove the `--clear-cache` flag of `flet build` and `flet debug` (deprecated in `0.86.0`). Use the `flet clean` command instead ([#6233](https://github.com/flet-dev/flet/issues/6233)) by @ndonkoHenri.
 
 ## 0.86.1